Use bluemonday for HTML sanitation, and don't use two different blackfridays.
